task: move StatusEnum to avoid circular imports
